[change] .gitignoreに機密情報ファイルaccounts.csvを追加
